Made achievement completion actually increment the point count and post on user profile fixed

+++ backend/models/achievement.go
@@ -11,7 +11,7 @@ type Achievement struct {
 	Description     string             `json:"description"`
 	CarbonReduction float32            `json:"carbon_reduction"`
 	Repeating       bool               `json:"repeating"`
-	Points          int8               `json:"points"`
+	Points          int32               `json:"points"`
 	AchievedBy      []bson.M           `json:"achieved_by"`
 	Details         bson.M             `json:"details"`
 	ImageURL		string				`json:"image_url"`
